+++ day_12/Cargo.toml
@@ -23,8 +23,12 @@ optional = true
 version = "0.20.3"
 optional = true
 
+[dependencies.bevy_mod_picking]
+version = "0.13.0"
+optional = true
+
 [features]
-bevy = ["dep:bevy", "dep:bevy_egui"]
+bevy = ["dep:bevy", "dep:bevy_egui", "dep:bevy_mod_picking"]
 
 [[bin]]
 name="hill-climbing"
