+++ day_12/src/map/constants.rs
@@ -1,0 +1,12 @@
+pub const CTA_FIND_PATH: &str = "Go!";
+pub const FIND_PATH_TOOLTIP: &str = "find the shortest path!";
+pub const CTA_TOGGLE_START: &str = "start";
+pub const TOGGLE_START_TOOLTIP: &str =
+  "Click on the green surface to place the start point.";
+pub const CTA_TOGGLE_END: &str = "end";
+pub const TOGGLE_END_TOOLTIP: &str =
+  "Click on the green surface to place the end point.";
+pub const NO_PATH_LABEL: &str = "No path found!";
+pub const NO_PATH_TITLE: &str = "No Path";
+pub const DESIRED_VIEW_WIDTH: f32 = 5.0;
+pub const DESIRED_VIEW_HEIGHT: f32 = 5.0;
