--- day_12/src/mapi/mod.rs
@@ -1,1 +1,0 @@
-
